[COUR-20] Implementation of basic course findFiltered

+++ src/main/java/it/amorabito/coursinho/model/dtos/CourseFilter.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 public class CourseFilter {
 
-    Integer minimumPrice;
-    Integer maximumPrice;
+    Double minimumPrice;
+    Double maximumPrice;
 }
